Rename the ListItemFactory
