+++ ui-components/app/components/file-list/index.js
@@ -1,8 +1,8 @@
 import Component from '../component.js';
 import FileItemComponent from '../file-item-component';
 import FolderItemComponent from '../folder-item-component';
-import FileItem from '../../models/list-items/file-item';
-import FolderItem from '../../models/list-items/folder-item';
+import FileItem from '../../models/file-system-objects/file-item';
+import FolderItem from '../../models/file-system-objects/folder-item';
 
 /**
  * The component for displaying the file list.
