+++ ui-components/app/models/file-system-objects/file-item/index.js
