+++ ui-components/app/models/file-system-objects/folder-item/index.js
