+++ ui-components/app/state/actions/get-files-action/index.js
@@ -2,7 +2,7 @@ import Action from '../action';
 import IsFileListLoadingMutator from '../../mutators/is-file-list-loading-mutator';
 import FileListMutator from '../../mutators/file-list-mutator';
 import FileListLoadingErrorMutator from '../../mutators/file-list-loading-error-mutator';
-import ListItemFactory from '../../../models/list-items/list-item-factory';
+import FileSystemObjectFactory from '../../../models/file-system-objects/file-system-object-factory';
 
 /**
  * The actions that gets files from the server.
@@ -24,7 +24,7 @@ export default class GetFilesAction extends Action {
     stateManager.mutate(new IsFileListLoadingMutator(true));
     try {
       const files = await apiService.getFiles(this._folderId);
-      const items = files.files.map((item) => ListItemFactory.createItem(item));
+      const items = files.files.map((item) => FileSystemObjectFactory.createItem(item));
       stateManager.mutate(new FileListMutator(items));
     } catch (e) {
       stateManager.mutate(new FileListLoadingErrorMutator(e));
