+++ ui-components/app/state/mutators/file-list-mutator/index.js
@@ -1,6 +1,6 @@
 import Mutator from '../mutator';
-import FileItem from '../../../models/list-items/file-item';
-import FolderItem from '../../../models/list-items/folder-item';
+import FileItem from '../../../models/file-system-objects/file-item';
+import FolderItem from '../../../models/file-system-objects/folder-item';
 
 /**
  * The mutator that saves the file list.
